+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Conjur CLI internally uses `rest_client_options` new to `conjur-api` 5.3.5.
+  [cyberark/conjur-cli#310](https://github.com/cyberark/conjur-cli/issues/310)
 
 ## [6.2.3] - 2020-12-22
 ### Fixed